fixed logic error in isFinished() of TurnDegrees and DriveDistance--was not handling negative degrees or negative distance correctly
i believe i fixed logic error in the isFinished() conditions to now handle turning in both directions (one neg, one pos) and driving in forward and ref (neg/pos); deleted old DriveDistance() renamed DriveDistance2() to DriveDistance()
